Add more application APIs
This also add the routes and service binding APIs required for complete e2e flow for a cloud foundry application

+++ examples/cf/app/main.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 	"log"
 	"os"
+	"time"
 
 	"github.com/IBM-Bluemix/bluemix-go/api/cf/cfv2"
 	"github.com/IBM-Bluemix/bluemix-go/session"
@@ -25,8 +26,11 @@ func main() {
 	var space string
 	flag.StringVar(&space, "space", "", "Bluemix Space")
 
+	var timeout time.Duration
+	flag.DurationVar(&timeout, "timeout", 120*time.Second, "Maximum time to wait for application to start")
+
 	var routeName string
-	flag.StringVar(&routeName, "routeName", "", "Bluemix app route")
+	flag.StringVar(&routeName, "route", "", "Bluemix app route")
 
 	var buildpack string
 	flag.StringVar(&buildpack, "buildpack", "", "Bluemix buildpack")
@@ -34,18 +38,21 @@ func main() {
 	var instance int
 	flag.IntVar(&instance, "instance", 2, "Bluemix App Instance")
 
+	var serviceInstanceName string
+	flag.StringVar(&serviceInstanceName, "svcname", "myservice", "Bluemix service instance name for the cloudantnosqldb offering")
+
 	var memory int
 	flag.IntVar(&memory, "memory", 128, "Bluemix app memory")
 
 	var diskQuota int
 	flag.IntVar(&diskQuota, "diskQuota", 512, "Bluemix app diskquota")
 
-	var async bool
-	flag.BoolVar(&async, "async", false, "For asynchronous and synchronous request. Defaults to false")
+	var clean bool
+	flag.BoolVar(&clean, "clean", false, "If set to true it will delete the application")
 
 	flag.Parse()
 
-	if name == "" || space == "" || buildpack == "" || org == "" || path == "" || routeName == "" {
+	if name == "" || space == "" || org == "" || path == "" || routeName == "" {
 		flag.Usage()
 		os.Exit(1)
 	}
@@ -77,10 +84,27 @@ func main() {
 	}
 	log.Println(myorg.GUID, myspace.GUID)
 
-	appAPI := client.Apps()
-	_, err = appAPI.Exists(myspace.GUID, name)
-
+	serviceOfferingAPI := client.ServiceOfferings()
+	myserviceOff, err := serviceOfferingAPI.FindByLabel("cloudantNoSQLDB")
 	if err != nil {
+		log.Fatal(err)
+	}
+	servicePlanAPI := client.ServicePlans()
+	plan, err := servicePlanAPI.GetServicePlan(myserviceOff.GUID, "Lite")
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	serviceInstanceAPI := client.ServiceInstances()
+	myService, err := serviceInstanceAPI.Create(serviceInstanceName, plan.GUID, myspace.GUID, nil, nil)
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	appAPI := client.Apps()
+	_, err = appAPI.FindByName(myspace.GUID, name)
+
+	if err == nil {
 		log.Fatal(err)
 	}
 
@@ -113,7 +137,12 @@ func main() {
 	}
 
 	if len(route) == 0 {
-		newroute, err := routeAPI.Create(routeName, domain.GUID, myspace.GUID)
+		req := cfv2.RouteRequest{
+			Host:       routeName,
+			DomainGUID: domain.GUID,
+			SpaceGUID:  myspace.GUID,
+		}
+		newroute, err := routeAPI.Create(req)
 		fmt.Println(newroute)
 		if err != nil {
 			log.Fatal(err)
@@ -138,10 +167,27 @@ func main() {
 		log.Fatal(err)
 	}
 
-	_, err = appAPI.Start(newapp.Metadata.GUID, async)
+	app, err := appAPI.Start(newapp.Metadata.GUID, timeout)
 	if err != nil {
 		log.Fatal(err)
 	}
+	log.Println("App status is", app.Entity.State)
+
+	sbAPI := client.ServiceBindings()
+
+	sb, err := sbAPI.Create(cfv2.ServiceBindingRequest{
+		ServiceInstanceGUID: myService.Metadata.GUID,
+		AppGUID:             app.Metadata.GUID,
+	})
+
+	if err != nil {
+		log.Fatal(err)
+	}
+	sbFields, err := sbAPI.Get(sb.Metadata.GUID)
+	if err != nil {
+		log.Fatal(err)
+	}
+	log.Println(*sbFields)
 
 	apps, err := appAPI.Get(newapp.Metadata.GUID)
 	fmt.Println(apps)
@@ -165,9 +211,11 @@ func main() {
 		log.Fatal(err)
 	}
 
-	err = appAPI.Delete(newapp.Metadata.GUID)
-	if err != nil {
-		log.Fatal(err)
+	if clean {
+		err = appAPI.Delete(newapp.Metadata.GUID)
+		if err != nil {
+			log.Fatal(err)
+		}
 	}
 
 }
